init changes for image upload
